Change callback behavior; allow stream updates
This commit changes callback behavior for Transcribers. onTranscriptionUpdated is now invoked with the current model generation, and onTranscriptionCommitted is invoked with the current running transcript (which is a sequence of generations that are periodically "committed" when the model input audio buffer has exceeded the length specified in MoonshineSettings).

This commit also allows StreamTranscriber to be reused with new streams, rather than requiring the developer to create a new StreamTranscriber each time they want to transcribe a new MediaStream. This is accessed through the new .attachStream(stream: MediaStream) method.