Update to invoke onTranscriptionUpdated at every pass
onTranscriptionUpdated used to be invoked only when speech was detected. There are cases, however, where we might want to know that the model is generating, but its outputs are empty. The behavior of this callback is changed so that it is invoked every time the model has new generation output (empty or not). The behavior of onTranscriptionCommitted is unchanged.