Handle multiple calls to load without re-downloading
Transcribers now store a static mapping of modelURLs to models. If a model of a given modelURL has been created already, Transcribers will reuse the same model instance without re-downloading it. On the other hand, if we create Transcribers with different modelURLs (i.e., model types), we will load the different models.